refactor(vanilla-extract): align meta structure with emotion/scss/tailwind | hj | #4542
Replace the loose single-entry model with discriminated token/utility leaves and a precise domain-keyed VanillaExtractMeta, add kindOf, reorganize into theme/ + utilities/ in scss domain order, drop deprecated from the catalog (derived from src barrels in the spec), and restructure grid into a nested-by-area tree.
