--- packages/styles/projects/vanilla-extract/vanillaExtractMeta/focus.meta.ts
@@ -1,17 +1,0 @@
-import { getFocusStyle, getFocusVisibleStyle } from '../src/focus/';
-import type { VanillaExtractMeta } from './meta.types';
-
-export const focusMeta: VanillaExtractMeta = {
-  getFocusVisibleStyle: {
-    name: 'getFocusVisibleStyle',
-    description: 'Applies a **focus-visible** style.',
-    value: getFocusVisibleStyle,
-  },
-  /** @deprecated since v4.0.0, will be removed with next major release. Use getFocusVisibleStyle instead. */
-  getFocusStyle: {
-    name: 'getFocusStyle',
-    description: 'deprecated since v4.0.0, will be removed with next major release. Use getFocusVisibleStyle instead.',
-    value: getFocusStyle,
-    deprecated: true,
-  },
-} as const;
