--- packages/styles/projects/vanilla-extract/vanillaExtractMeta/meta.types.ts
@@ -1,10 +1,0 @@
-export type VanillaExtractMetaEntry = {
-  name: string;
-  description: string;
-  value: string | number | readonly unknown[] | Record<string, unknown> | ((...args: any[]) => unknown);
-  deprecated?: boolean;
-};
-
-export type VanillaExtractMeta = {
-  [key: string]: VanillaExtractMetaEntry | { [key: string]: VanillaExtractMetaEntry };
-};
