--- packages/styles/projects/vanilla-extract/vanillaExtractMeta/typography.meta.ts
@@ -1,267 +1,0 @@
-import {
-  displayLargeStyle,
-  displayMediumStyle,
-  displaySmallStyle,
-  headingLargeStyle,
-  headingMediumStyle,
-  headingSmallStyle,
-  headingXLargeStyle,
-  headingXXLargeStyle,
-  proseHeading2XlStyle,
-  proseHeading2XsStyle,
-  proseHeading3XlStyle,
-  proseHeading4XlStyle,
-  proseHeading5XlStyle,
-  proseHeadingLgStyle,
-  proseHeadingMdStyle,
-  proseHeadingSmStyle,
-  proseHeadingXlStyle,
-  proseHeadingXsStyle,
-  proseText2XlStyle,
-  proseText2XsStyle,
-  proseText3XlStyle,
-  proseText4XlStyle,
-  proseText5XlStyle,
-  proseTextLgStyle,
-  proseTextMdStyle,
-  proseTextSmStyle,
-  proseTextXlStyle,
-  proseTextXsStyle,
-  textLargeStyle,
-  textMediumStyle,
-  textSmallStyle,
-  textXLargeStyle,
-  textXSmallStyle,
-  textXXSmallStyle,
-} from '../src/typography/';
-import type { VanillaExtractMeta } from './meta.types';
-
-export const typographyMeta: VanillaExtractMeta = {
-  proseHeading5XlStyle: {
-    name: 'proseHeading5XlStyle',
-    description:
-      'Applies the **5x-large** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeading5XlStyle,
-  },
-  proseText5XlStyle: {
-    name: 'proseText5XlStyle',
-    description:
-      'Applies the **5x-large** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseText5XlStyle,
-  },
-  proseHeading4XlStyle: {
-    name: 'proseHeading4XlStyle',
-    description:
-      'Applies the **4x-large** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeading4XlStyle,
-  },
-  proseText4XlStyle: {
-    name: 'proseText4XlStyle',
-    description:
-      'Applies the **4x-large** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseText4XlStyle,
-  },
-  proseHeading3XlStyle: {
-    name: 'proseHeading3XlStyle',
-    description:
-      'Applies the **3x-large** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeading3XlStyle,
-  },
-  proseText3XlStyle: {
-    name: 'proseText3XlStyle',
-    description:
-      'Applies the **3x-large** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseText3XlStyle,
-  },
-  proseHeading2XlStyle: {
-    name: 'proseHeading2XlStyle',
-    description:
-      'Applies the **2x-large** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeading2XlStyle,
-  },
-  proseText2XlStyle: {
-    name: 'proseText2XlStyle',
-    description:
-      'Applies the **2x-large** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseText2XlStyle,
-  },
-  proseHeadingXlStyle: {
-    name: 'proseHeadingXlStyle',
-    description:
-      'Applies the **x-large** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeadingXlStyle,
-  },
-  proseTextXlStyle: {
-    name: 'proseTextXlStyle',
-    description:
-      'Applies the **x-large** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseTextXlStyle,
-  },
-  proseHeadingLgStyle: {
-    name: 'proseHeadingLgStyle',
-    description:
-      'Applies the **large** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeadingLgStyle,
-  },
-  proseTextLgStyle: {
-    name: 'proseTextLgStyle',
-    description:
-      'Applies the **large** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseTextLgStyle,
-  },
-  proseHeadingMdStyle: {
-    name: 'proseHeadingMdStyle',
-    description:
-      'Applies the **medium** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeadingMdStyle,
-  },
-  proseTextMdStyle: {
-    name: 'proseTextMdStyle',
-    description:
-      'Applies the **medium** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseTextMdStyle,
-  },
-  proseHeadingSmStyle: {
-    name: 'proseHeadingSmStyle',
-    description:
-      'Applies the **small** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeadingSmStyle,
-  },
-  proseTextSmStyle: {
-    name: 'proseTextSmStyle',
-    description:
-      'Applies the **small** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseTextSmStyle,
-  },
-  proseHeadingXsStyle: {
-    name: 'proseHeadingXsStyle',
-    description:
-      'Applies the **x-small** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeadingXsStyle,
-  },
-  proseTextXsStyle: {
-    name: 'proseTextXsStyle',
-    description:
-      'Applies the **x-small** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseTextXsStyle,
-  },
-  proseHeading2XsStyle: {
-    name: 'proseHeading2XsStyle',
-    description:
-      'Applies the **2x-small** heading typography variant primarily to `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` tags.',
-    value: proseHeading2XsStyle,
-  },
-  proseText2XsStyle: {
-    name: 'proseText2XsStyle',
-    description:
-      'Applies the **2x-small** text typography variant primarily to `<p>`, `<ul>`, `<ol>`, `<blockquote>` tags.',
-    value: proseText2XsStyle,
-  },
-  display: {
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeading5XlStyle instead. */
-    displayLargeStyle: {
-      name: 'displayLargeStyle',
-      description:
-        'deprecated since v4.0.0, will be removed with next major release. Use proseHeading5XlStyle instead.',
-      value: displayLargeStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeading4XlStyle instead. */
-    displayMediumStyle: {
-      name: 'displayMediumStyle',
-      description:
-        'deprecated since v4.0.0, will be removed with next major release. Use proseHeading4XlStyle instead.',
-      value: displayMediumStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeading3XlStyle instead. */
-    displaySmallStyle: {
-      name: 'displaySmallStyle',
-      description:
-        'deprecated since v4.0.0, will be removed with next major release. Use proseHeading3XlStyle instead.',
-      value: displaySmallStyle,
-      deprecated: true,
-    },
-  },
-  heading: {
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeading2Xl instead. */
-    headingXXLargeStyle: {
-      name: 'headingXXLargeStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseHeading2Xl instead.',
-      value: headingXXLargeStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeadingXl instead. */
-    headingXLargeStyle: {
-      name: 'headingXLargeStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseHeadingXl instead.',
-      value: headingXLargeStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeadingLg instead. */
-    headingLargeStyle: {
-      name: 'headingLargeStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseHeadingLg instead.',
-      value: headingLargeStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeadingMd instead. */
-    headingMediumStyle: {
-      name: 'headingMediumStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseHeadingMd instead.',
-      value: headingMediumStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseHeadingSm instead. */
-    headingSmallStyle: {
-      name: 'headingSmallStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseHeadingSm instead.',
-      value: headingSmallStyle,
-      deprecated: true,
-    },
-  },
-  text: {
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseTextXl instead. */
-    textXLargeStyle: {
-      name: 'textXLargeStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseTextXl instead.',
-      value: textXLargeStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseTextLg instead. */
-    textLargeStyle: {
-      name: 'textLargeStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseTextLg instead.',
-      value: textLargeStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseTextMd instead. */
-    textMediumStyle: {
-      name: 'textMediumStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseTextMd instead.',
-      value: textMediumStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseTextSm instead. */
-    textSmallStyle: {
-      name: 'textSmallStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseTextSm instead.',
-      value: textSmallStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseTextXs instead. */
-    textXSmallStyle: {
-      name: 'textXSmallStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseTextXs instead.',
-      value: textXSmallStyle,
-      deprecated: true,
-    },
-    /** @deprecated since v4.0.0, will be removed with next major release. Use proseText2Xs instead. */
-    textXXSmallStyle: {
-      name: 'textXXSmallStyle',
-      description: 'deprecated since v4.0.0, will be removed with next major release. Use proseText2Xs instead.',
-      value: textXXSmallStyle,
-      deprecated: true,
-    },
-  },
-} as const;
