+++ packages/styles/projects/vanilla-extract/vanillaExtractMeta/utilities/focus.ts
@@ -1,0 +1,10 @@
+import { getFocusVisibleStyle } from '../../src/focus/';
+import type { VanillaExtractMeta } from '../types';
+
+export const focus = {
+  getFocusVisibleStyle: {
+    name: 'getFocusVisibleStyle',
+    description: 'Applies a **focus-visible** style.',
+    styles: getFocusVisibleStyle,
+  },
+} satisfies VanillaExtractMeta['focus'];
